Fixed an overlook in CClient.h
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1398 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CViewport.cpp
@@ -426,7 +426,7 @@ void CViewport::setSmooth(bool _b)
 {
 	bSmooth = _b; 
 	cSmoothVel = cSmoothAccel = CVec(0,0); 
-	curPos = tgtPos = CVec( WorldX, WorldY ); 
+	curPos = tgtPos = CVec( (float)WorldX, (float)WorldY ); 
 }
 
 // Constants that control behavior of smoothed viewport
